requirements.txt
streamlit
numpy
pandas
joblib
requests
python-dotenv
scikit-learn
thefuzz


